refactor(logging): split logging.h into atomic headers under rex/logging/
Extract the monolithic logging.h into focused sub-headers:
- logging/types.h: LogCategoryId, log:: constants, LogConfig, deprecated enum
- logging/api.h: function declarations and CVAR declarations
- logging/macros.h: REX_LOG_IMPL and all macro variants
- logging/assert.h: REX_FATAL_* and REX_ASSERT_* macros
- logging/format.h: rex::log::ptr/hex/boolean helpers
- logging/sink.h: LogEntry and LogCaptureSink (moved from log_capture.h)

logging.h becomes an umbrella header so existing includes are unchanged. Removed log_capture.h, migrated consumers to logging/sink.h directly.
